audience-affinity-flask-app/requirements.txt
# Install with: pip install -r requirements.txt

flask
pandas
faker
scikit-learn
plotly
joblib
google-generativeai

